% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdffromtable.R
\name{cdffromtable}
\alias{cdffromtable}
\title{CDF from Table Function}
\usage{
cdffromtable(values, T)
}
\arguments{
\item{values}{values=vector of values, on which histogram will be built}

\item{T}{T= a data frame with three columns describing a set of histograms.}
}
\description{
A function building and plotting CDF, with fitted lines for each branch, from range and frequency of bins given.
input
 T$min= left boundary, T$max=right boundary, T$frequency= bin frequency.
}
\examples{
cdffromtable()
}
\keyword{cats}
