% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESS.R
\name{DESS}
\alias{DESS}
\title{DESS Function}
\usage{
DESS(tree, values, kk, graph)
}
\arguments{
\item{tree}{tree= an hclust object created on "values"}

\item{values}{vector of values, on which histogram will be built}

\item{kk}{kk=Number of clusters at the tree-height the tree will be cut.}

\item{graph}{graph=Logical. Whether the histogram will be plotted or not.}
}
\description{
This function returns DESS values from all branches if tree is cut a fixed height.
}
\examples{
DESS()
}
\keyword{DESS}
