% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histbydess.R
\name{histbyDESS}
\alias{histbyDESS}
\title{Hist by DESS Function}
\usage{
histbyDESS(values, epsilon, L0, graph)
}
\description{
This function dynamically cuts a tree (not at fixed height) according to DESS threshold. (application of algorithm1)
inputs: 
values=vector of values, on which histogram will be built
epsilon: from the DESS criterion DESS < epsilon*(b_j-a_j)^2/3. Default value is 1 if missing
L0= minimum DESS values which is tolerated in the bins. Default is 0.1* tree height if missing.
Graph=Logical. Whether the histogram will be plotted or not. Default is FALSE.
Output:
 $table= a table containing maximum, minimum and frequency of each bin.
 $trees - a list all pruned branches from the main tree that constructed the histogram
 @param values
 @param epsilon
 @param L0
 @param graph
}
\examples{
histbyDESS()
}
\keyword{Hist}
