% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histfromtable.R
\name{histfromtable}
\alias{histfromtable}
\title{Histogram from Table Function}
\usage{
histfromtable(T)
}
\arguments{
\item{T}{T= a data frame with three columns describing a set of histograms. T$min= left boundary, T$max=right boundary, T$frequency= bin freuency.}
}
\description{
This function builds a histogram from range and frequency of bins given.
}
\examples{
histfromtable()
}
\keyword{cats}
