% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbydess.R
\name{plotDESS}
\alias{plotDESS}
\title{Plot DESS Function}
\usage{
plotDESS(values, M)
}
\arguments{
\item{M}{M=the output from the histbyDESS() function}
}
\description{
This function makes a plot using M, the output from the histbyDESS() function.
$all.dess= A list of all DESS values calculated from the bins
 $all.threshold = A list of (b_j-a_j)^2/3 values from the bins
}
\examples{
plotDESS()
}
\keyword{plot}
